refactor: 카카오 관련 로직 분리 (#15)
