+++ src/main/java/project/backend/business/auth/service/AuthService.java
@@ -1,32 +1,21 @@
 package project.backend.business.auth.service;
 
+import static project.backend.business.auth.service.oauth.KakaoApiService.BEARER;
+
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.Collections;
-import java.util.Map;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.client.HttpClientErrorException;
-import org.springframework.web.client.RestTemplate;
-import project.backend.common.auth.oauth.KakaoUserInfo;
+import project.backend.business.auth.service.oauth.KakaoApiService;
 import project.backend.dao.auth.repository.RefreshTokenRedisRepository;
 import project.backend.dao.user.entity.User;
-import project.backend.dao.user.repository.UserRepository;
 import project.backend.common.auth.token.RefreshToken;
 import project.backend.common.auth.token.TokenProvider;
 import project.backend.common.auth.token.TokenResponse;
@@ -37,27 +26,17 @@ import project.backend.common.auth.token.TokenResponse;
 @Transactional(readOnly = true)
 public class AuthService {
 
-  private final String BEARER = "Bearer ";
   private final TokenProvider tokenProvider;
-  private final UserRepository userRepository;
+  private final KakaoApiService kakaoApiService;
   private final RefreshTokenRedisRepository refreshTokenRedisRepository;
-
-  @Value("${spring.security.oauth2.client.registration.kakao.client-id}")
-  private String kakaoClientId;
-
-  @Value("${spring.security.oauth2.client.registration.kakao.client-secret}")
-  private String kakaoClientSecret;
-
-  @Value("${spring.security.oauth2.client.registration.kakao.redirect-uri}")
-  private String redirect_uri;
 
   @Value("${jwt.refresh_header}")
   private String refreshTokenHeader;
 
   @Transactional
   public TokenResponse kakaoLogin(String code) throws JsonProcessingException {
-    String token = getToken(code);
-    User user = getKakaoUser(token);
+    String token = kakaoApiService.getKakaoToken(code);
+    User user = kakaoApiService.getKakaoUser(token);
 
     TokenResponse tokenResponse = tokenProvider.createToken(
         String.valueOf(user.getId()), user.getEmail(), "USER");
@@ -67,91 +46,10 @@ public class AuthService {
     return tokenResponse;
   }
 
-  private void saveRefreshTokenOnRedis(User user, TokenResponse response) {
-    refreshTokenRedisRepository.save(RefreshToken.builder()
-                                                 .id(user.getId())
-                                                 .email(user.getEmail())
-                                                 .authorities(Collections.singleton(new SimpleGrantedAuthority("USER")))
-                                                 .refreshToken(response.getRefreshToken())
-                                                 .build());
-  }
-
-  private String getToken(final String code) throws JsonProcessingException {
-    // HTTP 헤더 생성
-    HttpHeaders headers = new HttpHeaders();
-    headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
-
-    // HTTP 바디 생성
-    MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
-    body.add("grant_type", "authorization_code");
-    body.add("client_id", kakaoClientId);
-    body.add("client_secret", kakaoClientSecret);
-    body.add("redirect_uri", redirect_uri);
-    body.add("code", code);
-
-    // HTTP 요청 보내기
-    HttpEntity<MultiValueMap<String, String>> kakaoTokenRequest = new HttpEntity<>(body, headers);
-
-    RestTemplate restTemplate = new RestTemplate();
-    ResponseEntity<String> response = restTemplate.exchange(
-        "https://kauth.kakao.com/oauth/token",
-        HttpMethod.POST,
-        kakaoTokenRequest,
-        String.class
-    );
-
-    String responseBody = response.getBody();
-    ObjectMapper objectMapper = new ObjectMapper();
-    JsonNode jsonNode = objectMapper.readTree(responseBody);
-
-    return jsonNode.get("access_token").asText();
-  }
-
-  @Transactional
-  public User getKakaoUser(String token) throws JsonProcessingException {
-    HttpHeaders headers = new HttpHeaders();
-    headers.add("Authorization", BEARER + token);
-    headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
-
-    HttpEntity<Void> kakaoProfileRequest = new HttpEntity<>(headers);
-
-    try {
-      RestTemplate restTemplate = new RestTemplate();
-      ResponseEntity<String> response = restTemplate.exchange(
-          "https://kapi.kakao.com/v2/user/me",
-          HttpMethod.POST,
-          kakaoProfileRequest,
-          String.class
-      );
-
-      String responseBody = response.getBody();
-      ObjectMapper objectMapper = new ObjectMapper();
-
-      // 응답을 Map<String, Object>로 파싱
-      Map<String, Object> attributes = objectMapper.readValue(responseBody, new TypeReference<>() {
-      });
-
-      // KakaoUserInfo 인스턴스 생성
-      KakaoUserInfo kakaoUserInfo = new KakaoUserInfo(attributes);
-
-      String email = kakaoUserInfo.getEmail();
-      String name = kakaoUserInfo.getName();
-      String profileImageUrl = kakaoUserInfo.getProfileImageUrl();
-
-      // 데이터베이스에서 사용자 조회 또는 생성
-      return userRepository.findByEmail(email)
-                           .orElseGet(() -> userRepository.save(
-                               User.createUser(email, name, profileImageUrl)
-                           ));
-    } catch (HttpClientErrorException e) {
-      throw new RuntimeException("카카오 사용자 정보를 가져오는데 실패했습니다.", e);
-    }
-  }
-
   public TokenResponse reissueAccessToken(HttpServletRequest request) {
     String refreshToken = extractRefreshToken(request).orElse(null);
 
-    // RefreshToken 이 유효하지 않을 경우 -> 재로그인 필요
+    // RefreshToken 이 유효하지 않을 경우
     if (!tokenProvider.validate(refreshToken) || !tokenProvider.validateExpired(refreshToken)) {
       throw new RuntimeException();
     }
@@ -176,8 +74,17 @@ public class AuthService {
     return tokenResponse;
   }
 
+  private void saveRefreshTokenOnRedis(User user, TokenResponse response) {
+    refreshTokenRedisRepository.save(RefreshToken.builder()
+                                                 .id(user.getId())
+                                                 .email(user.getEmail())
+                                                 .authorities(Collections.singleton(new SimpleGrantedAuthority("USER")))
+                                                 .refreshToken(response.getRefreshToken())
+                                                 .build());
+  }
+
   // RefreshToken 추출
-  public Optional<String> extractRefreshToken(HttpServletRequest request) {
+  private Optional<String> extractRefreshToken(HttpServletRequest request) {
     return Optional.ofNullable(request.getHeader(refreshTokenHeader))
                    .filter(refreshToken -> refreshToken.startsWith(BEARER))
                    .map(refreshToken -> refreshToken.replace(BEARER, ""));
